+++ src/components/portfolioSection.tsx
@@ -5,10 +5,12 @@ import './_3dCard.scss'
 import imgTest1 from '../assets/bg_coc_test.png'
 import imgTest2 from '../assets/logo_coc_test.png'
 import imgTest3 from '../assets/model_coc_test.webp'
+import SectionHeader from './sectionHeader'
 
 interface Test {
   label: string
   category: portfolioCategory
+  link: string
   images: {
     background: string
     logo: string
@@ -26,8 +28,9 @@ export default function PortfolioSection() {
 
   const projectItems = [
     {
-      label: 'Kiddo',
+      label: 'Projet Kiddo',
       category: portfolioCategory.WEB,
+      link: 'https://kiddo-front-la-passerelle.netlify.app/',
       images: {
         background: imgTest1,
         logo: imgTest2,
@@ -37,6 +40,27 @@ export default function PortfolioSection() {
     {
       label: 'Company Of Cube - Site',
       category: portfolioCategory.WEB,
+      link: 'https://companyofcube.fr/',
+      images: {
+        background: imgTest1,
+        logo: imgTest2,
+        character: imgTest3,
+      },
+    },
+    {
+      label: 'Fake Discussion',
+      category: portfolioCategory.WEB,
+      link: 'https://fake-discussion.netlify.app/',
+      images: {
+        background: imgTest1,
+        logo: imgTest2,
+        character: imgTest3,
+      },
+    },
+    {
+      label: 'Store BMS',
+      category: portfolioCategory.WEB,
+      link: 'http://www.store-bms.fr/',
       images: {
         background: imgTest1,
         logo: imgTest2,
@@ -46,6 +70,27 @@ export default function PortfolioSection() {
     {
       label: 'Company Of Cube - Jeu',
       category: portfolioCategory.JAVA,
+      link: 'https://companyofcube.fr/',
+      images: {
+        background: imgTest1,
+        logo: imgTest2,
+        character: imgTest3,
+      },
+    },
+    {
+      label: 'Installeur Minecraft moddé',
+      category: portfolioCategory.JAVA,
+      link: 'https://github.com/KinderrKill/Minecraft-Installer',
+      images: {
+        background: imgTest1,
+        logo: imgTest2,
+        character: imgTest3,
+      },
+    },
+    {
+      label: 'Minecraft JSON Generator',
+      category: portfolioCategory.JAVA,
+      link: 'https://github.com/KinderrKill/Minecraft-Json-Generator',
       images: {
         background: imgTest1,
         logo: imgTest2,
@@ -61,13 +106,11 @@ export default function PortfolioSection() {
   return (
     <>
       <div className='portfolio-section'>
-        <section className='portfolio-section__title flex flex-col justify-start items-center gap-10 pt-10 '>
-          <h2>Portfolio</h2>
-          <p className='text-center'>
-            Lorem Ipsum is simply dummy text of the printing and typesetting industry.<br></br>
-            Lorem Ipsum standard dummy text.
-          </p>
-        </section>
+        <SectionHeader id='portfolio' title='Portfolio'>
+          Lorem Ipsum is simply dummy text of the printing and typesetting industry.<br></br>
+          Lorem Ipsum standard dummy text.
+        </SectionHeader>
+
         <section className='flex flex-col items-center justify-center mt-16'>
           <article>
             <div className='flex gap-5'>
@@ -88,17 +131,19 @@ export default function PortfolioSection() {
               </button>
             </div>
           </article>
-          <article className='grid grid-cols-3'>
+          <article className='grid grid-cols-1 md:grid-cols-3'>
             {getListedItems().map((item: Test, index: number) => {
               return (
-                <a href='https://www.mythrillfiction.com/the-dark-rider' target='_blank'>
+                <a key={index} href={item.link} target='_blank' className='mb-16'>
                   <div className='card-3d'>
                     <div className='wrapper'>
                       <img src={item.images.background} className='cover-image' />
                     </div>
-                    <img src={item.images.logo} className='title' />
-                    {/* <span className='title text-center'>TEST TITLE</span> */}
                     <img src={item.images.character} className='character' />
+                    <span className='title text-center'>
+                      <img src={item.images.logo} />
+                      {item.label}
+                    </span>
                   </div>
                 </a>
               )