+++ src/pages/home.tsx
@@ -10,12 +10,15 @@ import ServiceSection from '../components/serviceSection'
 import PresentationSection from '../components/presentationSection'
 import PortfolioSection from '../components/portfolioSection'
 import ContactSection from '../components/contactSection'
+import TestSection from '../components/testSection'
 
 export default function HomePage() {
   return (
     <>
       <section className='section'>
-        <article className='content-panel'>
+        <article
+          className='content-panel w-full h-1/3 absolute top-1/2 left-0 -translate-y-1/2 px-5
+        lg:w-2/5 lg:left-52 lg:h-2/5'>
           <ConsoleCommand />
         </article>
         <article className='rocket-panel'>