+++ src/utils/constants.ts
@@ -1,4 +1,4 @@
-export const APP_VERSION = '0.5'
+export const APP_VERSION = '0.6'
 
 export const FETCH_METHOD = {
   GET_FULL_LIST: 'getFullList',